feat: Add session claims helper methods to check roles and permissions
